requirements.txt
python-hcl